requirements.txt
torch==2.4.0
biopython
mamba-ssm
causal-conv1d